+++ CHANGELOG.md
@@ -226,3 +226,4 @@ was removed or the new one was added to folder
 2.07
 - Simplified Text Rendering and naming conventions
 - Loading Single Texture is now safer and will return base texture if failed not black box
+- Ui Elements can have anonymous name when name parameter is empty at creation